20221107 #7
참석자 감정 분석 자료 수정
APP 미팅룸 종료 시 영상 파일 받기 - 강의 종료 데이터 수신
APP으로부터 사용자 확인 후 미팅룸 정보를 전달 수정

+++ src/main/java/com/eura/web/model/AnalysisMapper.java
@@ -7,5 +7,5 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface AnalysisMapper {
     //미팅 중 감정 데이터 저장
-    public Integer insertAnalysisData(LiveEmotionVO liveEmotionVO);
+    public Long insertAnalysisData(LiveEmotionVO liveEmotionVO);
 }
